Changed RgbColour and hex conversion to use Byte, not sure whay I used int.

+++ BlazorBuilds/Common/Models/AllSimpleTypes.cs
@@ -1,3 +1,3 @@
 ﻿namespace BlazorBuilds.Common.Models;
 
-public record RgbColour(int Red = 0, int Green = 0, int Blue = 0);
+public record RgbColour(byte Red = 0, byte Green = 0, byte Blue = 0);
